Add stubs for the robotics and drama club substories
SimpleStory subclasses can now receive notifications of when an option
is selected. This is demonstrated in schedule_conflict.gd, which adds
different stories to the list of those available. To make this work,
main_scene had to shuffle the list of available stories each draw
rather than only at the beginning; otherwise, new stories would be
added to the end and never show up.