Prevent presenter from making changes to the model
The SimpleStory is now responsible for updating the model, and the presenter
only shows the changes. We will need to put more thought into how to
show that the changes are affecting your own stats, e.g. by animating
the bonus from the main area down to the stats area, but that is
future work.