Revise main UI
The character name is now shown inside the character. The size is
approximated from a sketch; the actual size of the portrait likely
needs to change.  The colors of the buttons and text backgrounds don't
match the best of the UI sketches (e.g. revision 16, which this is
roughly based on).

The year display is based on values in World. It's a bit awkward right
now since a new one slides over the old one, but fixing that will
require more attention to the revised sliding UI.

Show year progress

IN A VERY INELEGANT WAY

Use observer pattern for turns change