Add ranks to attributes
An attribute gains points, and after some number of points
(currently 4), the attribute goes up in a rank. A star is
added to the corresponding attribute label.

This commit introduces the Attribute abstraction. Previously,
attributes were just integers, but now they are objects.
This made it easier to bind attributes to attribute labels.