Allow the player to skip the effect animations
This required a change in the way AttributeLabels work. Previously,
they updated themselves when attributes changed (observer
pattern). However, this doesn't work well with the idea of the player
being able to cancel animations through the presenter. Separating the
model from the view fixed this, although with the tradeoff that now
the presenter has to tell the character view to update itself. That is
the presenter's job, though, so it seems a fair tradeoff.