Allow stories to take place in different locations
The GameMap is like Cast: it holds all the possible
locations. Locations are simply textures for now. It did not seem
worthwhile to create a custom resource since there's only one datum
per location: the texture.

Integration tests ensure that the stories are properly using the
optional location field. If there isn't a location in a simple_story,
a random one is chosen.

Sample usage is shown in steven_dnd and engineering stories. These can
be changed later as needed; they are mostly for demonstration.