Fix locations on web build
The way we were loading location images worked locally but not in the web
build. This approach works, though it requires to have the locations listed
in the GameMap configuration. The integration test has been updated to
check whether the number of pngs in the locations folder matches the
registered locations in the GameMap in hopes that this prevents errors
later.