Clarify year transition
The routing of year information has changed so that it goes through
coroutines. This means that, unlike with signals, we can await on a
function's completion. The current approach to showing the year
transition is simple text with the Continue button. This is not great,
but it does solve a lot of problems: the player is already looking
there, and they know how to click "continue" to move on.