Rename Aubrey to Ethan
There was unnecessary confusion around the uncommon name 'Aubrey.'
We saw this particularly with the scout group. The name was only
ever a placeholder, and it stuck around too long. The character's
name is now 'Ethan'. Ethan was the most popular boy's name for
non-Hispanic Blacks in NYC in 2013 according to
https://www.nyc.gov/assets/doh/downloads/pdf/vs/baby-names-male2013.pdf