Replace command-line tool with scene-based
The command-line tool worked but was a bit tricky. It required extra
files and documentation to be able to run. Also, if it was run before
the project was opened in editor, it would spawn an infinite loop of
errors. Since the project had to be opened in-editor first anyway, the
new approach here uses an in-editor distribution report. Open the
StemercBalanceReport scene and it will show the current distribution
of STEMERC values across all SimpleStory subclasses. The report can be
regenerated by tapping the exported variable in the editor since there
is not a way to stick a button into the inspector.