Use multi-phase ending
This breaks down the player experience into a
series of phases, and at the end, shows them
a summary.

There is now a clearer separation between the
model of the epilogue, the text generation,
and the rendering of the view.

This change has near feature parity with the
old approach. The missing integration is the
canned end-stories, but I would like to lean
away from the old way of doing this; see
hacknplan for deatils.