Disable service worker
You may only have one service worker per page, but before this, we were
trying to load two: Godot's PWA script and coi-serviceworker. The
latter is required for this to work, so this commit will verify
whether or not the hypothesis is correct that we avoid bad
user experience by unifying on coi-serviceworker...at the cost
of sacrificing PWA features.