Release 1.0.1, including an upgrade to Godot 4.3
The new Web build in Godot 4.3 fixes the problems that prevented the
game from running on iOS and Mac devices. See
https://godotengine.org/article/progress-report-web-export-in-4-3/ for
some explanation from the Godot team.

The new build does not need the COI serviceworker. However, that
serviceworker is installed on any client that had previously loaded
the game. Hence, it has been replaced with a self-destroying one. The
fix for this is described at
https://medium.com/@nekrtemplar/self-destroying-serviceworker-73d62921d717

This also fixes an unused signal (not detected by static analysis
prior to 4.3) and adds the version label to the credits screen.