Remove local serving instructions
The feature to serve locally is already built in to Godot Engine.
When we started the project, this was not obvious, but we have
all been using the IDE feature to do this for some time. I am
just now getting around to cleaning out the old instructions
and script so that future readers don't get confused.